@@ -1,122 +1,146 @@
-const { Schema, model } = require('mongoose');
-const { sign } = require('jsonwebtoken');
-const { hash, compare } = require('bcryptjs');
-const { v4: uuidv4 } = require('uuid');
-const { toDate } = require('date-fns');
+import mongoose from 'mongoose';
+import { toDate } from 'date-fns';
+import uuid from 'uuid';
+import { nanoid } from 'nanoid';
+import bcryptjs from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+import constants from '../utils/constants/index.js';
+import config from '../config/index.js';
 
 const {
-  JWT_SECRET, JWT_ACCESS_TOKEN_EXPIRY, JWT_REFRESH_TOKEN_EXPIRY, JWT_ISSUER,
-} = require('../config').jwt;
+  JWT_ISSUER,
+  JWT_SECRET,
+  JWT_ACCESS_TOKEN_EXPIRY,
+  JWT_REFRESH_TOKEN_EXPIRY,
+} = config.jwt;
 
-const UserSchema = new Schema({
-  email: {
-    type: String,
-    unique: true,
-    index: true,
-    validate: {
-      validator(val) {
-        // eslint-disable-next-line no-useless-escape, max-len
-        const regex = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-        return regex.test(val);
-      },
-      message: ({
-        value,
-      }) => `${value} is not a valid email address`,
+const { Schema, model } = mongoose;
+
+const UserSchema = new Schema(
+  {
+    name: {
+      type: String,
+      minlength: [3, 'Name is too short!'],
+      maxlength: [27, 'Name is too long!'],
+      required: false,
+      trim: true,
     },
-    minlength: [5, 'Email is too short'],
-    maxlength: [120, 'Email is too long'],
-    trim: true,
-    required: [true, 'Email is required!'],
+    email: {
+      type: String,
+      unique: true,
+      index: true,
+      validate: {
+        validator(val) {
+          return constants.EMAIL_REGEX.test(val);
+        },
+        message: ({ value }) => `${value} is not a valid email address`,
+      },
+      minlength: [3, 'Email is too short!'],
+      maxlength: [120, 'Email is too long!'],
+      required: [true, 'Email is required!'],
+      trim: true,
+    },
+    username: {
+      type: String,
+      lowercase: true,
+      unique: true,
+      index: true,
+      minlength: [3, 'Username is too short!'],
+      maxlength: [120, 'Username is too long!'],
+      required: [true, 'Username is required!'],
+      trim: true,
+      default: uuid.v4(),
+    },
+    password: {
+      type: String,
+      minlength: [6, 'Password is too short!'],
+      maxlength: [240, 'Password is too long!'],
+      required: [true, 'Password is required!'],
+    },
+    mobile: {
+      type: String,
+      minlength: [6, 'Mobile number is too short!'],
+      maxlength: [15, 'Mobile number is too long!'],
+      required: false,
+      trim: true,
+    },
+    gender: {
+      type: String,
+      lowercase: true,
+      enum: ['female', 'male', 'others', ''],
+      default: '',
+      required: false,
+      trim: true,
+    },
+    dob: {
+      type: Date,
+    },
+    doj: {
+      type: Date,
+      default: toDate(new Date()),
+    },
+    isAdmin: {
+      type: Boolean,
+      default: false,
+      select: false,
+    },
+    isUserVerified: {
+      type: Boolean,
+      default: false,
+      select: false,
+    },
+    isUserPremium: {
+      type: Boolean,
+      default: false,
+      select: false,
+    },
+    isUserArchived: {
+      type: Boolean,
+      default: false,
+      select: false,
+    },
+    lastLogin: {
+      type: Date,
+    },
+    referralCode: {
+      type: String,
+      default: nanoid(),
+    },
+    refererCode: {
+      type: String,
+      required: false,
+    },
   },
-  username: {
-    type: String,
-    lowercase: true,
-    unique: true,
-    index: true,
-    minlength: [5, 'Username is too short'],
-    maxlength: [120, 'Username is too long'],
-    trim: true,
-    required: [true, 'Username is required!'],
-    default: uuidv4(),
+  {
+    strict: true,
+    timestamps: true,
   },
-  password: {
-    type: String,
-    minlength: [8, 'Password is too short'],
-    trim: true,
-    required: [true, 'password is required!'],
-  },
-  gender: {
-    type: String,
-    trim: true,
-    lowercase: true,
-    enum: ['female', 'male', 'others', 'null'],
-    default: 'null',
-  },
-  dob: {
-    type: Date,
-  },
-  doj: {
-    type: Date,
-    default: toDate(new Date()),
-  },
-  isAdmin: {
-    type: Boolean,
-    default: false,
-    select: false,
-  },
-  isUserVerified: {
-    type: Boolean,
-    default: false,
-    select: false,
-  },
-  isUserPremium: {
-    type: Boolean,
-    default: false,
-    select: false,
-  },
-  isUserArchived: {
-    type: Boolean,
-    default: false,
-    select: false,
-  },
-  lastLogin: {
-    type: Date,
-  },
-  referralCode: {
-    type: String,
-  },
-  refererCode: {
-    type: String,
-  },
-}, {
-  timestamps: true,
-  strict: true,
-});
+);
 
 UserSchema.pre('save', async function hashPassword(next) {
   try {
     if (this.isModified('password') || this.isNew) {
-      const hashedPassword = await hash(this.password, 12);
+      const hashedPassword = await bcryptjs.hash(this.password, 12);
       this.password = hashedPassword;
       return next();
     }
     return true;
   } catch (err) {
     console.error(err);
-    throw err;
+    throw new Error(err);
   }
 });
 
 UserSchema.methods.comparePassword = async function comparePassword(password) {
   try {
-    return await compare(password, this.password);
+    return await bcryptjs.compare(password, this.password);
   } catch (err) {
     console.error(err);
-    throw err;
+    throw new Error(err);
   }
 };
 
-UserSchema.methods.genRefreshToken = async function genRefreshToken() {
+UserSchema.methods.genRefreshToken = function genRefreshToken() {
   try {
     const payload = {
       id: this._id,
@@ -124,7 +148,7 @@ UserSchema.methods.genRefreshToken = async function genRefreshToken() {
       email: this.email,
       isAdmin: this.isAdmin,
     };
-    const token = sign(payload, JWT_SECRET, {
+    const token = jwt.sign(payload, JWT_SECRET, {
       algorithm: 'HS384',
       expiresIn: JWT_REFRESH_TOKEN_EXPIRY,
       issuer: JWT_ISSUER,
@@ -132,11 +156,11 @@ UserSchema.methods.genRefreshToken = async function genRefreshToken() {
     return token;
   } catch (err) {
     console.error(err);
-    throw err;
+    throw new Error(err);
   }
 };
 
-UserSchema.methods.genAccessToken = async function genAccessToken() {
+UserSchema.methods.genAccessToken = function genAccessToken() {
   try {
     const payload = {
       id: this._id,
@@ -144,7 +168,7 @@ UserSchema.methods.genAccessToken = async function genAccessToken() {
       email: this.email,
       isAdmin: this.isAdmin,
     };
-    const token = sign(payload, JWT_SECRET, {
+    const token = jwt.sign(payload, JWT_SECRET, {
       algorithm: 'HS384',
       expiresIn: JWT_ACCESS_TOKEN_EXPIRY,
       issuer: JWT_ISSUER,
@@ -152,11 +176,11 @@ UserSchema.methods.genAccessToken = async function genAccessToken() {
     return token;
   } catch (err) {
     console.error(err);
-    throw err;
+    throw new Error(err);
   }
 };
 
-UserSchema.methods.formattedUserObj = async function formattedUserObj() {
+UserSchema.methods.formattedUserObj = function formattedUserObj() {
   const obj = this.toObject({
     virtuals: true,
   });
@@ -171,10 +195,10 @@ UserSchema.methods.formattedUserObj = async function formattedUserObj() {
     __v,
     ...rest
   } = obj;
-  const formattedObj = {
+  const userObj = {
     ...rest,
   };
-  return formattedObj;
+  return userObj;
 };
 
-module.exports = model('User', UserSchema);
+export default model('User', UserSchema);
